feat(payments): move currency cloud pagination function to common package
