+++ internal/connectors/plugins/public/currencycloud/accounts.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/formancehq/payments/internal/connectors/plugins/public/currencycloud/client"
 	"github.com/formancehq/payments/internal/models"
+	"github.com/formancehq/payments/internal/utils/pagination"
 )
 
 type accountsState struct {
@@ -50,7 +51,7 @@ func (p Plugin) fetchNextAccounts(ctx context.Context, req models.FetchNextAccou
 		}
 
 		needMore := true
-		needMore, hasMore, accounts = shouldFetchMore(accounts, nextPage, req.PageSize)
+		needMore, hasMore, accounts = pagination.ShouldFetchMore(accounts, nextPage, req.PageSize)
 		if !needMore {
 			break
 		}
